Fixes an issue where cards that error would get loaded late and cover the commander
This introduces a bunch of Wait.time and coroutine stuff so that the system waits (without blocking) for the API calls to resolve before moving forward with the Scryfall request.

This also introduces a ton of improvements to the batching, and makes batching into groups of 20, which still lags a bunch but feels slightly better because of the visual indicator of what batch you're on.